requirements/testing.txt
coverage[toml]
pytest
pytest-cov
pytest-django